+++ gcp/gcp.go
@@ -2,104 +2,66 @@
 // Use of this source code is governed by a MIT license found in the LICENSE file.
 
 // Package gcp provides customization for application runs on GCP.
-//
-// It integrates with following GCP services:
-//   - [Cloud Logging]
-//   - [Cloud Error Reporting]
-//   - [Cloud Trace]
-//   - [Cloud Monitoring]
-//   - [Cloud Profiler]
-//
-// [Cloud Logging]: https://cloud.google.com/logging
-// [Cloud Error Reporting]: https://cloud.google.com/error-reporting
-// [Cloud Trace]: https://cloud.google.com/trace
-// [Cloud Monitoring]: https://cloud.google.com/monitoring
-// [Cloud Profiler]: https://cloud.google.com/profiler
 package gcp
 
 import (
 	"context"
-	"fmt"
-	"os"
+	"log/slog"
 
-	"cloud.google.com/go/compute/metadata"
 	"github.com/nil-go/sloth/gcp"
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc"
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
-	"go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/resource"
-	"go.opentelemetry.io/otel/sdk/trace"
+	"github.com/nil-go/sloth/otel"
+	"github.com/nil-go/sloth/rate"
+	"github.com/nil-go/sloth/sampling"
+	"go.opentelemetry.io/otel/trace"
 
 	"github.com/nil-go/nilgo/gcp/profiler"
 )
 
-// Args provides the [nilgo.Run] args with given options for application runs on GCP.
-func Args(opts ...Option) ([]any, error) { //nolint:cyclop,funlen
+// Logger returns a slog.Logger integrated with [Cloud Logging] and [Cloud Error Reporting].
+//
+// [Cloud Logging]: https://cloud.google.com/logging
+// [Cloud Error Reporting]: https://cloud.google.com/error-reporting
+func Logger(opts ...Option) *slog.Logger {
 	option := options{}
-	for _, opt := range opts {
-		opt(&option)
-	}
+	option.apply(opts)
+	logOpts := append(
+		[]gcp.Option{
+			gcp.WithErrorReporting(option.service, option.version),
+			gcp.WithTrace(option.project),
+		},
+		option.logOpts...,
+	)
 
-	// Get service and version from Google Cloud Run environment variables.
-	if option.service == "" {
-		option.service = os.Getenv("K_SERVICE")
-	}
-	if option.version == "" {
-		option.version = os.Getenv("K_REVISION")
-	}
-	if option.project == "" {
-		option.project, _ = metadata.ProjectID()
-	}
+	handler := gcp.New(logOpts...)
+	handler = rate.New(handler)
+	handler = sampling.New(handler,
+		func(ctx context.Context) bool {
+			sc := trace.SpanContextFromContext(ctx)
 
-	var appOpts []any
-	if option.logOpts != nil {
-		option.logOpts = append([]gcp.Option{gcp.WithErrorReporting(option.service, option.version)}, option.logOpts...)
-		if option.traceOpts != nil {
-			option.logOpts = append([]gcp.Option{gcp.WithTrace(option.project)}, option.logOpts...)
-		}
-		appOpts = append(appOpts, gcp.New(option.logOpts...))
-	}
+			return !sc.IsValid() || sc.IsSampled()
+		},
+	)
+	handler = otel.New(handler)
 
-	res := resource.Default()
-	ctx := context.Background()
-	if option.traceOpts != nil {
-		exporter, err := otlptracegrpc.New(ctx, option.traceOpts...)
-		if err != nil {
-			return nil, fmt.Errorf("create otlp trace exporter: %w", err)
-		}
-		appOpts = append(appOpts,
-			trace.NewTracerProvider(
-				trace.WithBatcher(exporter),
-				trace.WithResource(res),
-				trace.WithSampler(trace.ParentBased(trace.NeverSample())),
-			),
-		)
-	}
-	if option.metricOpts != nil {
-		exporter, err := otlpmetricgrpc.New(ctx, option.metricOpts...)
-		if err != nil {
-			return nil, fmt.Errorf("create otlp metric exporter: %w", err)
-		}
+	return slog.New(handler)
+}
 
-		appOpts = append(appOpts,
-			metric.NewMeterProvider(
-				metric.WithReader(metric.NewPeriodicReader(exporter)),
-				metric.WithResource(res),
-			),
-		)
-	}
+// Profiler returns a function to start [Cloud Profiler].
+// It requires the following IAM roles:
+//   - roles/cloudprofiler.agent
+//
+// [Cloud Profiler]: https://cloud.google.com/profiler
+func Profiler(opts ...Option) func(context.Context) error {
+	option := options{}
+	option.apply(opts)
+	profilerOpts := append(
+		[]profiler.Option{
+			profiler.WithProject(option.project),
+			profiler.WithService(option.service),
+			profiler.WithVersion(option.version),
+		},
+		option.profilerOpts...,
+	)
 
-	if option.profilerOpts != nil {
-		option.profilerOpts = append(
-			[]profiler.Option{
-				profiler.WithProject(option.project),
-				profiler.WithService(option.service),
-				profiler.WithVersion(option.version),
-			},
-			option.profilerOpts...,
-		)
-		appOpts = append(appOpts, profiler.Run(option.profilerOpts...))
-	}
-
-	return appOpts, nil
+	return profiler.Run(profilerOpts...)
 }